developing gui
i created a separate .py file for testing the GUI. i watched some youtube tutorials to learn how to use CustomerTKInter to make a clean, modern layout for my app. i set it up so that there are several "pages" that you can switch between using a function. the button uses the function as the command