dynamic variables and setting up the remaining frames
today i figured out how to gather user input in an entry widget and use the text to display a username. it was hard because at first i wasnt using a tk.stringvar(). once i learned how to use them, i was able to print a welcome message: "Hello [name-from-input]"